fix: 코드 래빗 리뷰 반영(#44)
